# refine sequence enumerate op