pyramid dnn infer 0.186378ms => 0.174904ms, speedup 1.1%
op: 27.7244ms => 25.5843ms , speedup 1.1%
Before 
```
I0319 06:59:58.893966  7319 helper.h:270] ====== batch_size: 1, repeat: 1000, threads: 1, thread id: 0, latency: 0.186378ms, fps: 5365.44 ======
...
Event                                Calls       Total       Min.        Max.        Ave.        Ratio.
thread0::sequence_enumerate          12000       27.7244     0.002041    0.007117    0.00231037  0.136199
```

After
```
I0319 06:51:45.478598 29663 helper.h:270] ====== batch_size: 1, repeat: 1000, threads: 1, thread id: 0, latency: 0.174904ms, fps: 5717.41 ======
thread0::sequence_enumerate          12000       25.5843     0.001927    0.113019    0.00213202  0.131972
```